+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Version 7.0.0-dev
 - BC break: removed (virtual) method Job::onAfterExecute
 - BC break: removed BaseAttribute
 - fixed failure line in JUnit reports
+- BC break: replaced method IResultsFormatter::render with outputResults
 
 Version 6.1.0
 - added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api